corrected translation vars

+++ install.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 MODS=$1
 cp -f bin/${MODS}.dll GameData/QuickMods/${MODS}/Plugins/
-rm -rf Lang/*~
-cp -rf Lang/ GameData/QuickMods/${MODS}/
+if [ -e Lang ]; then
+        rm -rf Lang/*~
+        cp -rf Lang/ GameData/QuickMods/${MODS}/
+fi
 cp -f README.md GameData/QuickMods/${MODS}/
 cp -f COPYING GameData/QuickMods/${MODS}/
 cp -f ${MODS}.version GameData/QuickMods/${MODS}/
